+++ src/main/java/org/kumoricon/site/attendee/search/byname/AttendeeSearchCheckInView.java
@@ -55,6 +55,17 @@ public class AttendeeSearchCheckInView extends CheckInView implements View {
     }
 
     @Override
+    protected void btnEditClicked() {
+        if (attendee.getOrder() != null) {
+            navigateTo(SearchByNameView.VIEW_NAME + "/" + attendee.getOrder().getOrderId() + "/" + attendeeId);
+        } else {
+            // TODO: if staff get an orderID set, remove this
+            navigateTo(SearchByNameView.VIEW_NAME + "/" + "staff" + "/" + attendeeId);
+        }
+
+    }
+
+    @Override
     public void close() {
         if (attendeeId != null) {
             navigateTo(SearchByNameView.VIEW_NAME + "/" + searchString + "/" + attendeeId);